dgmr/requirements.txt
torch
antialiased_cnns
pytorch_msssim
numpy
torchvision>=0.11.0
pytorch_lightning
einops
huggingface_hub
datasets
TensorFlow
tensorrt
wandb
netcdf4
pysftp
xarray
opencv-python
pytest
python-dotenv